It seems to be faster to make a new vec each time when reading
This kinda makes some sense, most blocks will be full, so on average, we
would be making a new vec of the same length, and copying into it
anyway. This doesn't apply as much to compression (we hope we only have
to copy a smallish part of a block), and it seems to slightly hurt (or
be in the noise) to migrate compression.