Handle creating temp directories when destination is a file
When applesauce is passed a file on a different device than /tmp, it
should try to make a tempdir in the _parent_ of the path, not somehow
inside the file.

Also, handle errors in adding a tempdir destination: we already handle
the case where we don't have a common directory for a temp file by
creating a tempfile in the same directory.