feat: refuse to break hard links
Because we write to a temporary file, then atomically replace the
original file, doing so on a file that has hard link(s) would break
those links (the other files would not be updated, and any future writes
to this or those links would no longer update the other links).

In order to avoid the posibility that could cause breakage, cowardly
refuse to compress(or decompress) files with hardlinks.